requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
icalendar==4.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
schedule==1.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typeguard==2.13.3
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
